requirements.txt
atomicwrites==1.4.0
attrs==20.3.0
blinker==1.4
click==7.1.2
colorama==0.4.4
dnspython==2.1.0
email-validator==1.1.2
Flask==1.1.2
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Moment==0.11.0
flask-mongoengine==1.0.0
Flask-WTF==0.14.3
greenlet==1.0.0
gunicorn==20.1.0
idna==3.1
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mongoengine==0.23.0
packaging==20.9
pluggy==0.13.1
py==1.10.0
PyJWT==2.0.1
pyngrok==5.0.5
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.17.0
PyYAML==5.4.1
six==1.15.0
toml==0.10.2
Werkzeug==1.0.1
WTForms==2.3.3
